% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cks2018.R
\docType{package}
\name{CKS2018-package}
\alias{CKS2018}
\alias{CKS2018-package}
\title{Calculate Table 1 for C K S 2018}
\description{
This packages calculates tables in C K S 2018.
}
\author{
\strong{Maintainer}: John Kolassa \email{kolassa@stat.rutgers.edu}

Authors:
\itemize{
  \item Harold Sackrowitz \email{sackrowi@stat.rutgers.edu}
  \item Maryanne Sackrowitz
}

}
