% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbic.R
\name{gbic}
\alias{gbic}
\title{Apply generalized BIC model selection.}
\usage{
gbic(Treat, Cntrl, ssq, nsam, Cons, cstr = NULL, verbose = FALSE)
}
\arguments{
\item{Treat}{A vector of treatment means.}

\item{Cntrl}{A single control mean}

\item{ssq}{Sum of squares within treatments and control, divided by the number of observations making up each mean.}

\item{nsam}{Number of independent observations making up the means.}

\item{Cons}{Generalized BIC constants.}

\item{cstr}{Work constants, regenerated if not given.}

\item{verbose}{Logical flag indicating whether intermediate results are printed.}
}
\value{
Models selected.
}
\description{
Apply generalized BIC model selection.
}
\examples{
#Original single control multiple (3) treatments example of Dunnett (1955).
control<-c(55,47,48)
treatment<-array(c(55,55,50,64,49,44,64,52,41),c(3,3))
Cntrl<-mean(control)
Treat<-apply(treatment,1,mean)
ssq<-(sum((Cntrl-control)^2)+sum(apply(treatment,2,"-",Treat)^2))/length(control)
Cons<-c(1000000,20.5,8,0.00)
gbic(Treat,Cntrl,ssq,length(control),Cons)
}
