% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCstep.R
\name{makeCstep}
\alias{makeCstep}
\title{Generate constants in stepwise constraints.}
\usage{
makeCstep(Numval, gamma, nsam, maxp = NA)
}
\arguments{
\item{Numval}{Number of parameters.}

\item{gamma}{The constant related to stepwise constraints.}

\item{nsam}{Number of independent observations making up the means.}

\item{maxp}{Maximum number of nonzero parameters to consider.}
}
\value{
Stepwise constraints.
}
\description{
Generate constants in stepwise constraints.
}
