% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1.R
\name{table1}
\alias{table1}
\title{Generate a table like Table 1 of Cohen, Kolassa, and Sackrowitz}
\usage{
table1(Cons = c(rep(1e+06, 8), 20.5, 8, 0), nsam = 14, gamma = 0.05,
  Nrep = 5000, printme = FALSE, setseed = FALSE, BIC = NA, maxp = NA,
  irange = NULL)
}
\arguments{
\item{Cons}{A vector of constants for the generalized BIC.}

\item{nsam}{A sample size for each of the means investigated.}

\item{gamma}{The constant related to stepwise constraints.}

\item{Nrep}{Number of Monte Carlo samples for size and power calculations.}

\item{printme}{Logical flag indicating whether intermediate results are printed.}

\item{setseed}{Logical flag indicating whether random seed will be.}

\item{BIC}{Integer. if not NA, number of categories to apply traditonal BIC.}

\item{maxp}{Integer. if not NA, maximum numbers of parameters in the model.}

\item{irange}{range for means for table construction.}
}
\value{
A table of the format of Table 1.
}
\description{
Generate a table like Table 1 of Cohen, Kolassa, and Sackrowitz
}
\examples{
#Call to table1 regenerates Table 1.
\donttest{
tab1<-round(table1(setseed=TRUE,maxp=2)[c(6:4,13,12,10,14:16,19:24),],2)
tab2a<-table1(printme=TRUE,setseed=TRUE,BIC=10)
tab2b<-table1(printme=TRUE,setseed=TRUE,BIC=10,maxp=2)
tab2<-cbind(tab2a[,1:4],tab2b[,3:4])[c(5,4,12:10,14:17,19:21,23),]
}
#A version of Table 1 based on fewer samples.
tab1<-table1(printme=TRUE,setseed=TRUE,Nrep=2)
}
