% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1line.R
\name{table1line}
\alias{table1line}
\title{Generate a single line of a table like Table 1 of Cohen, Kolassa, and Sackrowitz}
\usage{
table1line(Mu, Cons, Cstep, nsam, Nrep = 10000, printme = FALSE)
}
\arguments{
\item{Mu}{A vector of sample means for simulation.}

\item{Cons}{A vector of constants for the generalized BIC.}

\item{Cstep}{A vector of constants for stepwise selection.}

\item{nsam}{A sample size for each of the means investigated.}

\item{Nrep}{Number of Monte Carlo samples for size and power calculations.}

\item{printme}{Logical flag indicating whether intermediate results are printed to the screen.}
}
\value{
A line of a table of the format of Table 1.
}
\description{
Generate a single line of a table like Table 1 of Cohen, Kolassa, and Sackrowitz
}
