% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateerrors.R
\name{updateerrors}
\alias{updateerrors}
\title{Update counts of errors to track behavior of various tests, called within table1line}
\usage{
updateerrors(errors, FR, Nmu)
}
\arguments{
\item{errors}{A matrix of errors}

\item{FR}{output from gbic or steptest}

\item{Nmu}{number of comparisons.}
}
\value{
the updated error matrix
}
\description{
Update counts of errors to track behavior of various tests, called within table1line
}
